requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.1
    # via black
aiosignal==1.3.1
    # via aiohttp
aiosqlite==0.17.0
    # via tortoise-orm
altgraph==0.17.4
    # via pyinstaller
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   httpx
    #   starlette
    #   watchfiles
appdirs==1.4.4
    # via FakeSMS (pyproject.toml)
attrs==23.1.0
    # via aiohttp
black==23.12.0
    # via FakeSMS (pyproject.toml)
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   black
    #   uvicorn
fastapi==0.105.0
    # via FakeSMS (pyproject.toml)
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.2
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.25.2
    # via FakeSMS (pyproject.toml)
idna==3.6
    # via
    #   anyio
    #   httpx
    #   yarl
iso8601==1.1.0
    # via tortoise-orm
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
packaging==23.2
    # via
    #   black
    #   pyinstaller
pathspec==0.12.1
    # via black
platformdirs==4.1.0
    # via black
pydantic==2.5.2
    # via fastapi
pydantic-core==2.14.5
    # via pydantic
pyinstaller==6.3.0
    # via FakeSMS (pyproject.toml)
pyinstaller-hooks-contrib==2023.11
    # via pyinstaller
pypika-tortoise==0.1.6
    # via tortoise-orm
python-dotenv==1.0.0
    # via uvicorn
pytz==2023.3.post1
    # via tortoise-orm
pyyaml==6.0.1
    # via uvicorn
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
starlette==0.27.0
    # via fastapi
tomlkit==0.12.3
    # via FakeSMS (pyproject.toml)
tortoise-orm==0.20.0
    # via FakeSMS (pyproject.toml)
typing-extensions==4.9.0
    # via
    #   aiosqlite
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn[standard]==0.24.0.post1
    # via FakeSMS (pyproject.toml)
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
